+++ .rubocop.yml
@@ -7,6 +7,4 @@ Metrics/BlockLength:
   Enabled: false
 Layout/LineLength:
   Enabled: false
-Style/StringLiterals
-  Enabled: false
